libceph,rbd: client based replication support
This, of couse, is only for testing. There is no any attempt to failover
or any other smart things. The Grand Idea is to test possible throughput
speed up in case of client-based replication. Basically this patch
works only if OSDs do not fail.

To map rbd device with client-based replication (when all OP_WRITE or
OP_WRITEFULL requests replicated by the client, not by the OSD) can be
done with the following command:

  rbd device map rbd/fio_test -o client_based_replication

  * corresponding patch should be applied for rbd tool.

Signed-off-by: Roman Penyaev <rpenyaev@suse.de>